test(payroll): golden tests pays — cas moteur génériques centralisés, valeurs légales sourcées (Closes #1938)
