+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+- **test(payroll): golden tests pays — cas moteur génériques centralisés, valeurs légales sourcées (Closes #1938).** Les suites golden BF/ML/GA/CG dupliquaient à l'identique des cas dérivés du code (prorata `base × jours/22`, taux horaire `base / MONTHLY_HOURS`, fin de contrat `notice/30 × base` + `severanceMonthsPerYear() × années × base`) qui re-testaient le moteur générique et verrouillaient le défaut générique `1,0 mois/an` comme valeur légale. (1) Nouveau `GoldenEngineGenericTest` : mécaniques moteur centralisées (2 proratas, taux horaire + majoration 1,15, mécanique préavis/indemnité) — explicitement marquées NON légales. (2) Suites pays nettoyées : seules les valeurs légales sourcées restent (ex. BF : seuil HS 40 h/semaine §6 + préavis 30 j §7 ; ML/GA/CG : préavis 30 j §7) ; les assertions de mécanique (taux horaire, severance générique) sont retirées. (3) Corrections de comptage : les « 14 cas calculés à la main » annoncés en #1829/#1824 étaient gonflés par ces cas dérivés — les cas réellement sourcés sont désormais dénombrables via `payroll-golden-report.sh`.
+
+- **fix(infra/test/payroll): main rouge — 3 régressions débloquées (Closes #1962, #2024, #1930).** PR combinée : (1) migration tenant `add_rules_version_to_payroll_records` renumérotée `000013` → `000009` (collision basename avec `add_original_slip_id_to_pay_slips`, garde #1962) ; (2) `TaxRateChangeLogTest` — accolade de classe mal placée retirée (ParseError → suite de tests entière illisible, Backend Coverage rouge, régression #2024) ; (3) `IslamicCalendarService::resolveForPayroll()` — `$holiday['confirmed'] ?? false` retiré (offset garanti par le shape, erreur PHPStan Strict + Modules Architecture, régression #1930). Chaque fix individuel (#2062/#2063) héritait des autres pannes de main → aucune PR ne passait ses checks → blocage en chaîne ; la PR combinée casse le cycle.
+
 - feat(tooling): Spec Kit v0.16.3 — Spec-Driven Development intégré. Constitution `.specify/constitution.md`, 5 presets de conformité paie (leopardo-payroll / leopardo-multitenancy / leopardo-dz / leopardo-cemac / leopardo-cedeao), bundle agent, skills GitHub Copilot. Règle anti-doublon documentée dans AGENTS.md.
 - **fix(i18n): checksums `versions.json` régénérés — garde CI `I18N Enterprise` verte (Closes #2041).** Les PRs i18n parallèles (#1987/#2001) avaient régénéré `shared/i18n/versions/versions.json` avec des checksums divergents ; la résolution de conflit « prendre main » a laissé des checksums obsolètes → `node shared/i18n/validators/validate.js` (exécuté par `.github/workflows/i18n-enterprise.yml`, étape « Validate shared catalogs ») échouait sur les 4 locales : `checksum mismatch in versions.json`. Régénération via `node shared/i18n/sync/sync-mobile.js` (source canonique `shared/i18n/locales/*.json`, SHA-256 stable) → `I18N_VALIDATION_OK (4 locales)`. Aucun autre fichier touché (ARB/web/backend déjà alignés). Note process pour les agents : toute modification de catalogue i18n doit régénérer `versions.json` via le sync tooling, jamais à la main.
