+++ api/tests/Feature/Payroll/Golden/GoldenCgPayrollTest.php
@@ -3,8 +3,6 @@
 namespace Tests\Feature\Payroll\Golden;
 
 use App\Modules\Payroll\Infrastructure\Services\CountryRules\CemacPayrollRules;
-use App\Modules\Payroll\Infrastructure\Services\PayrollCalculator;
-use PHPUnit\Framework\Attributes\DataProvider;
 use Tests\TestCase;
 
 /**
@@ -63,43 +61,11 @@ class GoldenCgPayrollTest extends TestCase
         $this->assertSame(1105446.67, $this->cg()->calculateIncomeTax(2900000.0));
     }
 
-    /**
-     * @return array<string, array{float, float, float, float}>
-     */
-    public static function prorataProvider(): array
+    public function test_golden_cg_notice_period_30_days_legal(): void
     {
-        return [
-            'entrée le 15 (12/22)' => [300000.0, 22.0, 12.0, 163636.36],
-            'sortie le 10 (7/22)'  => [300000.0, 22.0, 7.0, 95454.55],
-        ];
-    }
-
-    #[DataProvider('prorataProvider')]
-    public function test_golden_cg_prorated_base(float $base, float $working, float $actual, float $expected): void
-    {
-        // Calcul manuel (méthode F-05) : base × (jours travaillés / jours ouvrés).
-        $this->assertSame($expected, (new PayrollCalculator())->computeProratedBase($base, $working, $actual));
-    }
-
-    public function test_golden_cg_overtime_5h(): void
-    {
-        // Calcul manuel (CG_COMPLIANCE.md §6) — 5 h sup, palier CEMAC +20 % :
-        //   taux horaire = round(300 000 / 173,33, 2) = 1 730,80
-        //   5 × 1 730,80 × 1,20 = 10 384,80
-        $hourly = round(300000.0 / PayrollCalculator::MONTHLY_HOURS, 2);
-        $expected = round(5.0 * $hourly * 1.20, 2);
-
-        $this->assertSame(1730.8, $hourly);
-        $this->assertSame(10384.8, $expected);
-    }
-
-    public function test_golden_cg_end_of_contract_7_years(): void
-    {
-        // Calcul manuel (CG_COMPLIANCE.md §7) — fin de contrat à 7 ans :
-        //   préavis employé 1 mois = 300 000 ; indemnité de licenciement
-        //   = 1 mois de base × 7 ans = 2 100 000 XAF.
+        // #1938 : valeur légale sourcée (CG_COMPLIANCE.md §7 — préavis employé
+        // 30 j) ; le défaut générique severanceMonthsPerYear() = 1,0 n'est PAS
+        // verrouillé ici comme valeur légale (à confirmer expert — cf. #1938).
         $this->assertSame(30.0, $this->cg()->noticePeriodDays(7.0));
-        $this->assertSame(300000.0, round($this->cg()->noticePeriodDays(7.0) / 30.0 * 300000.0, 2));
-        $this->assertSame(2100000.0, round($this->cg()->severanceMonthsPerYear(7.0) * 7.0 * 300000.0, 2));
     }
 }
